src/requirements.txt
# ============================================================================
# TaskPilot PoC - Python Dependencies
# Версия: 1.0
# ============================================================================

# ----------------------------------------------------------------------------
# Web Framework
# ----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
python-multipart==0.0.6

# ----------------------------------------------------------------------------
# Database (PostgreSQL + SQLAlchemy)
# ----------------------------------------------------------------------------
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ----------------------------------------------------------------------------
# Async & Queue (Celery + Redis)
# ----------------------------------------------------------------------------
celery==5.3.6
redis==5.0.1
kombu==5.3.5

# ----------------------------------------------------------------------------
# Security (JWT + Password Hashing)
# ----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ----------------------------------------------------------------------------
# LLM & AI (Mistral + Embeddings + FAISS)
# ----------------------------------------------------------------------------
mistralai==0.1.8
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# ----------------------------------------------------------------------------
# Observability (Metrics + Logging)
# ----------------------------------------------------------------------------
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# ----------------------------------------------------------------------------
# Validation & Settings
# ----------------------------------------------------------------------------
pydantic==2.5.3
pydantic-settings==2.1.0

# ----------------------------------------------------------------------------
# HTTP Client
# ----------------------------------------------------------------------------
httpx>=0.25.2,<0.26.0
tenacity==8.2.3

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
python-dotenv==1.0.0

# ----------------------------------------------------------------------------
# Security (JWT + Rate Limiting)
# ----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
slowapi==0.1.9

# ----------------------------------------------------------------------------
# Observability (OpenTelemetry + Prometheus)
# ----------------------------------------------------------------------------
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==0.42b0
prometheus-client==0.19.0
psutil==5.9.7

# ----------------------------------------------------------------------------
# LLM Metrics
# ----------------------------------------------------------------------------
# (уже есть mistralai, добавим только утилиты)
tiktoken==0.5.2
